Rename modules to be less redundant (asteroid_frontend -> frontend)
