+++ asteroid/lex.py
@@ -6,7 +6,7 @@
 
 import re
 
-from asteroid.asteroid_state import state
+from asteroid.state import state
 
 # table that specifies the token value and type for keywords
 keywords = {
