+++ asteroid/modules/dataframe.ast
@@ -17,7 +17,7 @@ function read_csv
 ------------------------------------------------------------------
 with (file,header_row) do escape
 "
-from asteroid.asteroid_state import state
+from asteroid.state import state
 import pandas
 
 this_val = state.symbol_table.lookup_sym('this')
@@ -45,7 +45,7 @@ with ix do return escape
 "
 global __retval__
 import numpy
-from asteroid.asteroid_state import state
+from asteroid.state import state
 
 this_val = state.symbol_table.lookup_sym('this')
 ix_val = state.symbol_table.lookup_sym('ix')
@@ -84,7 +84,7 @@ function col
 with ix do return escape
 "
 global __retval__
-from asteroid.asteroid_state import state
+from asteroid.state import state
 import numpy
 this_val = state.symbol_table.lookup_sym('this')
 ix_val = state.symbol_table.lookup_sym('ix')
@@ -123,7 +123,7 @@ function nrow
 with none do return escape
 "
 global __retval__
-from asteroid.asteroid_state import state
+from asteroid.state import state
 this_val = state.symbol_table.lookup_sym('this')
 
 # destructure object
@@ -143,7 +143,7 @@ function ncol
 with none do return escape
 "
 global __retval__
-from asteroid.asteroid_state import state
+from asteroid.state import state
 this_val = state.symbol_table.lookup_sym('this')
 
 # destructure object
