+++ asteroid/symtab.py
